add FORCE_VERSION to allow building always-install images
note this also allows building empty images, which might not be what
most people want

+++ scripts/versions.sh
@@ -97,7 +97,8 @@ gen_newversion() {
 
 	# if no version changed, clean up and fail script to avoid
 	# downloading the rest of the image
-	if cmp -s /etc/sw-versions "$SCRIPTSDIR/sw-versions.merged"; then
+	if cmp -s /etc/sw-versions "$SCRIPTSDIR/sw-versions.merged" \
+	    && ! grep -q "#VERSION_FORCE" "$SWDESC"; then
 		rm -rf "$SCRIPTSDIR"
 		error "Nothing to do -- failing on purpose to save bandwidth"
 	fi
